Refactor query normalization logic
- Moved `normalizeQueryTransformationOptions` to the `appDataCaseConversion` utility module for better organization.
- Updated `useAppData` to utilize the new normalization function, ensuring consistent handling of query options.
- Removed redundant normalization logic from `useAppData`, streamlining the code.
- Added comments to clarify the purpose of normalization for query options.
