+++ frontend/src/AppBuilder/_stores/slices/dataQuerySlice.js
@@ -5,7 +5,7 @@ import _, { isEmpty, throttle } from 'lodash';
 import { toast } from 'react-hot-toast';
 import { isQueryRunnable } from '@/_helpers/utils';
 import { replaceQueryOptionsEntityReferencesWithIds } from '@/AppBuilder/_stores/utils';
-import { normalizeQueryTransformationOptions } from '@/AppBuilder/_hooks/useAppData';
+import { normalizeQueryTransformationOptions } from '@/AppBuilder/_stores/utils/appDataCaseConversion';
 import { clearQueryRerunTimer } from '@/AppBuilder/_stores/slices/componentsSlice';
 
 const initialState = {
